+++ frontend/src/features/map_planning/routes/MapWrapper.tsx
@@ -1,8 +1,10 @@
 import { useRef, useEffect } from 'react';
+import { useTranslation } from 'react-i18next';
 import { ShepherdTour } from 'react-shepherd';
 import { LayerType, LayerDto } from '@/api_types/definitions';
 import { createAPI } from '@/config/axios';
 import { PolygonGeometry } from '@/features/map_planning/types/PolygonTypes';
+import { errorToastGrouped } from '@/features/toasts/groupedToast';
 import { useSafeAuth } from '@/hooks/useSafeAuth';
 import { EditorMap } from '../components/EditorMap';
 import {
@@ -13,6 +15,7 @@ import {
   usePlantLayer,
 } from '../hooks/mapEditorHookApi';
 import { useTourStatus } from '../hooks/tourHookApi';
+import useGetTimeLineData from '../hooks/useGetTimelineData';
 import { useMapId } from '../hooks/useMapId';
 import useMapStore from '../store/MapStore';
 import { handleRemoteAction } from '../store/RemoteActions';
@@ -26,14 +29,48 @@ function getDefaultLayer(layerType: LayerType, layers?: LayerDto[]) {
   return layers?.find((l) => l.type_ === layerType && !l.is_alternative);
 }
 
+const useInitializeTimeline = (mapId: number) => {
+  const timeLineVisibleYears = useMapStore((state) => state.untrackedState.timeLineVisibleYears);
+  const timeLineEvents = useGetTimeLineData(
+    mapId,
+    timeLineVisibleYears.from,
+    timeLineVisibleYears.to,
+  );
+
+  useEffect(() => {
+    if (timeLineEvents) {
+      useMapStore.setState((state) => ({
+        ...state,
+        untrackedState: {
+          ...state.untrackedState,
+          timeLineEvents: {
+            daily: timeLineEvents.daily,
+            monthly: timeLineEvents.monthly,
+            yearly: timeLineEvents.yearly,
+          },
+        },
+      }));
+    }
+  }, [timeLineEvents]);
+
+  return timeLineEvents;
+};
+
 /**
  * Hook that initializes the map by fetching all map data, layers and layer elements.
  */
 function useInitializeMap() {
   const mapId = useMapId();
-  const { data: layers } = useGetLayers(mapId);
   const { data: map } = useMap(mapId);
+  const { t } = useTranslation(['layers']);
+  const { data: layers, error } = useGetLayers(mapId);
 
+  if (error) {
+    console.error(error);
+    errorToastGrouped(t('layers:error_fetching_layers'), { autoClose: false });
+  }
+
+  const timeLineData = useInitializeTimeline(mapId);
   useEffect(() => {
     if (!layers) return;
 
@@ -89,7 +126,7 @@ function useInitializeMap() {
     }));
   }, [map]);
 
-  const isLoading = !layers || !map;
+  const isLoading = !layers || !map || !timeLineData;
 
   if (isLoading) {
     return null;